[SPOT-292][REFACTOR] 게시글 및 댓글 서비스 클래스 정리
