--- src/main/java/com/example/spot/post/application/LikedPostQueryService.java
@@ -1,8 +1,0 @@
-package com.example.spot.post.application;
-
-@Deprecated
-public interface LikedPostQueryService {
-    long countByPostId(Long postId);
-
-    boolean existsByMemberIdAndPostId(Long postId);
-}
