+++ src/main/java/com/example/spot/post/application/query/GetLikedPostUseCase.java
@@ -1,0 +1,8 @@
+package com.example.spot.post.application.query;
+
+public interface GetLikedPostUseCase {
+
+    long countByPostId(Long postId);
+
+    boolean existsByMemberIdAndPostId(Long postId);
+}
